+++ index.d.ts
@@ -82,10 +82,12 @@ declare module 'match-iz' {
   export function endsWith(text: string): TPredicate
   export function includes(content: any): TPredicate
 
-  export function not(value: any): TPredicate
-  export function allOf(...these: any[]): TPredicate
-  export function anyOf(...these: any[]): any[]
-  export function includedIn(...these: any): any[]
+  export function not(value: TPattern): TPredicate
+  export function allOf(...these: TPattern[]): TPredicate
+  export function anyOf(...these: TPattern[]): TPredicate
+  export function firstOf(...these: TPattern[]): TPredicate
+  export function lastOf(...these: TPattern[]): TPredicate
+  export function includedIn(...these: TPattern): TPredicate
   export function hasOwn(...props: string[]): TPredicate
   export function instanceOf(constructor: any): TPredicate